AP-2140 date_client_told_incident
