+++ config/routes.rb
@@ -203,7 +203,7 @@ Rails.application.routes.draw do
       resource :identify_types_of_outgoing, only: %i[show update]
       resource :opponent, only: %i[show update]
       resource :opponent_name, only: %i[show update]
-      resource :date_client_told_incident, only: %i[show update]
+      resource :date_client_told_incident, only: %i[show update], controller: 'application_merits_task/date_client_told_incidents'
       resource :proceedings_before_the_court, only: %i[show update]
       resources :success_likely, only: %i[index create]
       resource :success_prospects, only: %i[show update]
